move StringFieldAccessor
