+++ celements-model/src/main/java/com/celements/model/access/DefaultModelAccessFacade.java
@@ -36,9 +36,9 @@ import com.celements.model.classes.ClassIdentity;
 import com.celements.model.classes.fields.ClassField;
 import com.celements.model.context.ModelContext;
 import com.celements.model.field.FieldAccessor;
+import com.celements.model.field.StringFieldAccessor;
 import com.celements.model.field.XObjectFieldAccessor;
-import com.celements.model.field.internal.StringFieldAccessor;
-import com.celements.model.field.internal.XObjectStringFieldAccessor;
+import com.celements.model.field.XObjectStringFieldAccessor;
 import com.celements.model.object.ObjectFetcher;
 import com.celements.model.object.xwiki.XWikiObjectEditor;
 import com.celements.model.object.xwiki.XWikiObjectFetcher;
