+++ celements-model/src/main/resources/META-INF/components.txt
@@ -24,7 +24,7 @@ com.celements.model.util.DefaultModelUtils
 com.celements.model.access.DefaultXWikiDocumentCreator
 com.celements.model.access.StoreModelAccessStrategy
 com.celements.model.field.XObjectFieldAccessor
-com.celements.model.field.internal.XObjectInternalFieldAccessor
+com.celements.model.field.XObjectStringFieldAccessor
 com.celements.convert.bean.BeanFieldAccessor
 com.celements.convert.bean.XObjectBeanConverter
 com.celements.convert.bean.BeanXObjectConverter
